setup pipeline for datasets

+++ dags/scripts/fetch_datasets.py
@@ -8,6 +8,8 @@ def fetch_dataset():
     base_dir = "./DustiniaDelixia_Groceria"
     output_dir = "/opt/airflow/data_lake"
 
+    os.makedirs(output_dir, exist_ok=True)
+
     csv_files = [
         "customers.csv",
         "geolocation.csv",
@@ -16,77 +18,23 @@ def fetch_dataset():
         "orders.csv",
         "sellers.csv",
     ]
-    
-    for order in orders:
-        order_id = str(order.get("order_id"))
-        user_id = str(order.get("user_id"))
-        order_number = int(order.get("order_number", 0))
-        order_dow = int(order.get("order_dow", 0))
-        order_hour_of_day = int(
-            order.get("order_hour_of_day", 0)
-        )
-        days_since_prior_order = (
-            int(order.get("days_since_prior_order"))
-            if order.get("days_since_prior_order") is not None
-            else 0
-        )
-        eval_set = str(order.get("eval_set"))
-        order_date = datetime.now().strftime("%Y-%m-%d")
-        products = order.get("products", [])
 
-        for product in products:
-            parsed_data.append({
-                "order_id": order_id,
-                "user_id": user_id,
-                "order_number": order_number,
-                "order_dow": order_dow,
-                "order_hour_of_day": order_hour_of_day,
-                "days_since_prior_order":
-                    days_since_prior_order,
-                "eval_set": eval_set,
-                "product_id": str(
-                    product.get("product_id")
-                ),
-                "product_name": str(
-                    product.get("product_name")
-                ),
-                "aisle_id": int(
-                    product.get("aisle_id", 0)
-                ),
-                "aisle": str(
-                    product.get("aisle")
-                ),
-                "department_id": int(
-                    product.get("department_id", 0)
-                ),
-                "department": str(
-                    product.get("department")
-                ),
-                "add_to_cart_order": int(
-                    product.get("add_to_cart_order", 0)
-                ),
-                "reordered": int(
-                    product.get("reordered", 0)
-                ),
-                "order_date": order_date,
-                "updated_at": datetime.now().strftime(
-                    "%Y-%m-%d %H:%M:%S"
-                ),
-                "ingested_at": datetime.now().strftime(
-                    "%Y-%m-%d %H:%M:%S"
-                )
-            })
+    for file in csv_files:
+        source_path = os.path.join(base_dir, file)
+        if not os.path.exists(source_path):
+            print(f"{file} tidak ditemukan")
+            continue
 
-    df = pd.DataFrame(parsed_data)
+        df = pd.read_csv(source_path)
 
-    os.makedirs(output_dir, exist_ok=True)
-    output_path = (
-        f"{output_dir}/orders_"
-        f"{datetime.now().strftime('%Y%m%d_%H%M%S')}.parquet"
-    )
-    df.to_parquet(output_path, index=False)
+        parquet_name = file.replace(".csv", ".parquet")
+        output_path = os.path.join(output_dir, parquet_name)
 
-    print(f"✅ Berhasil menyimpan {len(df)} rows")
+        df.to_parquet(output_path, index=False)
+        print(f"✅ {file} -> {output_path}")
+
+    print("🎉 Semua dataset berhasil dimasukkan ke Data Lake")
 
 if __name__ == "__main__":
-    fetch_orders()
+    fetch_dataset()
+
